+++ packages/styleguide/.storybook/components/Navigation/index.ts
@@ -1,3 +1,5 @@
 export * from './NavigationProvider';
 export * from './types';
 export * from './utils';
+export * from './TableOfContents';
+export * from './BreadCrumbs';
