--- packages/styleguide/.storybook/components/styles.ts
@@ -1,8 +1,0 @@
-import { compose, system } from '@codecademy/gamut-system';
-import { values } from 'lodash';
-
-export const { properties, variant, ...groups } = system.create({});
-
-export const allProps = compose(...values(groups));
-
-export const { color, typography, spacing, layout } = groups;
